images check before delete

+++ app/Http/Livewire/Admin/AdminProductComponent.php
@@ -9,14 +9,17 @@ class AdminProductComponent extends Component
 {
     public function deleteProduct($id){
         $product = Product::find($id);
-        $images = explode(",", $product->images);
 
-        foreach ($images as $image ){
-            if ($image){
-                if(file_exists(public_path('assets/images/products/'. $image))){
-                    unlink(public_path('assets/images/products/'. $image));
-                }           
-            }             
+        if($product->images){
+            $images = explode(",", $product->images);
+
+            foreach ($images as $image ){
+                if ($image){
+                    if(file_exists(public_path('assets/images/products/'. $image))){
+                        unlink(public_path('assets/images/products/'. $image));
+                    }           
+                }             
+            }
         }
 
         if(file_exists(public_path('assets/images/products/'. $product->image))){
